optimize dequantize code structure
