+++ bitblas/ops/general_matmul/tilelang/dense/gemv_simt.py
@@ -82,9 +82,9 @@ class GemvFineGrainSIMTScheduler(MatmulSIMTBaseScheduler):
             "reduce_thread must be provided currently, as related bitblas.gpu.gemv.GEMV"
             "sch_outer_reduction_with_config is not implemented")
 
-        M, N, K = self.M, self.N, self.K
-        if not isinstance(M, int):
-            M = tvm.te.var("m")
+        M = self.maybe_dynamic(self.M, "m")
+        N, K = self.N, self.K
+        assert isinstance(N, int) and isinstance(K, int), "Do not support dynamic N and K Currently"
 
         in_dtype, out_dtype, accum_dtype = (
             self.in_dtype,
